+++ libs/ui/src/admin/ArtApprovalModal/ArtApprovalForm.tsx
@@ -1,38 +1,24 @@
-import { FC, memo, useState } from 'react'
+import { FC, memo } from 'react'
 
 import {
   Avatar,
-  Button,
-  ButtonGroup,
   HStack,
   Text,
-  Menu,
-  MenuButton,
-  MenuItem,
-  MenuList,
   Modal,
   ModalBody,
   ModalCloseButton,
   ModalContent,
   ModalOverlay,
   Stack,
-  Textarea,
   Box,
-  VStack,
   Image,
   Center,
   Flex,
   SimpleGrid,
-  IconButton,
 } from '@chakra-ui/react'
 import { Splide, SplideSlide } from '@splidejs/react-splide'
-import {
-  HiDotsVertical,
-  HiOutlineCheck,
-  HiOutlineX,
-  HiPencil,
-} from 'react-icons/hi'
 
+import { ArtFeedbackForm } from './ArtFeedbackForm'
 import { ArtImageProps, ArtApprovalFormTypes } from './types'
 
 const ArtImage: FC<ArtImageProps> = memo(({ image, alt }) => {
@@ -62,18 +48,6 @@ export const ArtApprovalForm: FC<ArtApprovalFormTypes> = ({
   artArtistName,
   isEdit,
 }) => {
-  const [feedback, setFeedback] = useState('')
-
-  const handleReject = () => {
-    onReject(artId, editorId, feedback)
-  }
-
-  const handleApprove = () => {
-    onApprove(artId, editorId, feedback)
-  }
-  const handleDelete = () => {
-    onDelete(artId)
-  }
   return (
     <Box>
       <Modal onClose={onClose} size={['1088px']} isOpen={isOpen}>
@@ -118,89 +92,17 @@ export const ArtApprovalForm: FC<ArtApprovalFormTypes> = ({
                     <Text>{artDescription}</Text>
                   </Flex>
                 </Stack>
-
-                <VStack align={'start'} w={'full'}>
-                  <Flex align="start" justify={'start'} w="full">
-                    <Text color={'black'} fontWeight={'bold'}>
-                      Give Feedback
-                    </Text>
-                  </Flex>
-
-                  <HStack w="full">
-                    {/* avatar*/}
-                    <Stack align={'start'} flex={1} textAlign="center">
-                      <Avatar size="sm" src={editorAvatar} name={editorName} />
-                    </Stack>
-                    <VStack w="full">
-                      {/* text area, button group*/}
-                      <Stack w={'full'}>
-                        {/* text area*/}
-                        <Textarea
-                          isRequired
-                          onChange={e => setFeedback(e.target.value)}
-                          placeholder={'type your comment here'}
-                        ></Textarea>
-                      </Stack>
-                      <Stack w={'full'}>
-                        {/*button group*/}
-                        <ButtonGroup
-                          alignItems="space-between"
-                          w={'full'}
-                          justifyItems="flex-end"
-                          spacing={4}
-                        >
-                          <Button
-                            onClick={handleReject}
-                            colorScheme="red"
-                            w="full"
-                            leftIcon={<HiOutlineX />}
-                          >
-                            Reject
-                          </Button>
-                          <Button
-                            onClick={handleApprove}
-                            colorScheme="green"
-                            w="full"
-                            leftIcon={<HiOutlineCheck />}
-                          >
-                            Approve
-                          </Button>
-                          <Menu>
-                            <MenuButton
-                              aria-label="Open art menu"
-                              as={IconButton}
-                              icon={<HiDotsVertical />}
-                              colorScheme="primary"
-                            ></MenuButton>
-                            <MenuList minWidth={32} minH={20}>
-                              <MenuItem
-                                as={Button}
-                                onClick={() => isEdit}
-                                variant="ghost"
-                                colorScheme="red"
-                                icon={<HiPencil />}
-                                _hover={{ bg: 'blue.400', color: 'white' }}
-                              >
-                                Edit{' '}
-                              </MenuItem>
-
-                              <MenuItem
-                                as={Button}
-                                onClick={handleDelete}
-                                variant="ghost"
-                                colorScheme="red"
-                                icon={<HiOutlineX />}
-                                _hover={{ bg: 'red', color: 'white' }}
-                              >
-                                Delete
-                              </MenuItem>
-                            </MenuList>
-                          </Menu>
-                        </ButtonGroup>
-                      </Stack>
-                    </VStack>
-                  </HStack>
-                </VStack>
+                {/*feedback ================================= */}
+                <ArtFeedbackForm
+                  onReject={onReject}
+                  onApprove={onApprove}
+                  onDelete={onDelete}
+                  artId={artId}
+                  editorId={editorId}
+                  editorAvatar={editorAvatar}
+                  editorName={editorName}
+                  isEdit={isEdit}
+                />
               </Stack>
             </SimpleGrid>
           </ModalBody>
